requirements.txt
keyboard
tk
tkinterdnd2
pywin32
icoextract
pillow